move TriggerState to its own file
