+++ src/QuartzNET-DynamoDB/QuartzNET-DynamoDB.csproj
@@ -87,6 +87,7 @@
     <Compile Include="DataModel\DynamoTriggerGroup.cs" />
     <Compile Include="DataModel\DynamoJobGroup.cs" />
     <Compile Include="DataModel\DynamoCalendar.cs" />
+    <Compile Include="DataModel\DynamoTriggerState.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="job_scheduling_data_2_0.xsd">
